LayoutTests/css3/transform-computed-style-001-expected.txt

Transformed elements:

matrix(1.5 0 0 0.75 -50 25)

matrix(0.866025 0.5 -0.5 0.866025 63.3975 -36.6025)

matrix(1 0 0 1 50 80)

matrix(0.787846 -0.138919 0.138919 0.787846 17.3235 85.1072)

matrix(1 0 0 1 0 0)
